teste finalizado

+++ database/migrations/2023_09_01_031321_create_dentistas_especialidades_table.php
@@ -13,11 +13,12 @@ class CreateDentistasEspecialidadesTable extends Migration
      */
     public function up()
     {
+       
         Schema::create('dentistas_especialidades', function (Blueprint $table) {
             $table->unsignedBigInteger('dentista_id');
-            $table->unsignedBigInteger('especialidade_id');
+            $table->unsignedInteger('especialidade_id'); // Alterado para unsignedInteger
             $table->timestamps();
-
+    
             $table->foreign('dentista_id')->references('id')->on('dentistas')->onDelete('cascade');
             $table->foreign('especialidade_id')->references('id')->on('especialidades')->onDelete('cascade');
         });
